Add: more parsing functions prototype in parser.h
With the view to conform to the grammar of every statement, server function is introduced.
For function _expr(void) it does not take any parameter and return a Binary tree node's address.
For function _expr_tail(BTNode*) it takes one parameter, which represent the node parsed according to the grammar,
and return a Binary tree node's address.

The parsing function still remain in light of some still being called.